AI-BASED-KNOWLEDGE-GRAPH-BUILDER-main/requirements.txt
chainlit
python-dotenv
pandas
numpy
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
spacy
neo4j
faiss-cpu
pypdf
protobuf